+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 * Refactor record control hooks. Refs [UILD-744].
 * Refactor resource processing and preview loading. Refs [UILD-744].
 * Refactor resource loading. Refs [UILD-816].
+* Fix default profile type persistence across edit form and profile settings. Fixes [UILD-820].
 
 [UILD-744]:https://folio-org.atlassian.net/browse/UILD-744
 [UILD-816]:https://folio-org.atlassian.net/browse/UILD-816
+[UILD-820]:https://folio-org.atlassian.net/browse/UILD-820
 
 ## 2.0.3 (2026-05-19)
 * Fix Reset search option is missing after using Advanced Search. Fixes [UILD-812].